Fetch selling plans and show within product
